backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.9
pgvector==0.2.5
llama-index==0.10.30
llama-index-llms-gemini==0.1.6
llama-index-embeddings-gemini==0.1.4
google-generativeai==0.4.1
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pypdf==4.0.0
python-docx==1.1.0
python-dotenv==1.0.0
websockets==12.0
python-multipart==0.0.6
alembic==1.13.1
asyncio==3.4.3
pybreaker==2.0.1